+++ Develappers.BillomatNet.Tests/IntegrationTests/UnitServiceIntegrationTests.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Develappers.BillomatNet.Queries;
@@ -7,9 +11,13 @@ using Xunit;
 
 namespace Develappers.BillomatNet.Tests.IntegrationTests
 {
-    [Trait(TraitNames.Category, CategoryNames.IntegrationTest)]
-    public class UnitServiceIntegrationTests
+    [Trait(Traits.Category, Traits.Categories.IntegrationTest)]
+    public class UnitServiceIntegrationTests : IntegrationTestBase<UnitService>
     {
+        public UnitServiceIntegrationTests() : base(c => new UnitService(c))
+        {
+        }
+
         [Fact]
         public async Task GetListOfUnits()
         {